Better fix for involunatry cycling back.
The problem was that happened if the user pressed the back up button up
after having selected a top-level item, it would go back to where it
game from, creating a involunatry cycle.

we're settnig the clicked top level item to null, to prevent triggering
the observable multiple times when the navController naviates back to
it, thus breaking the cycle that would have happened.